refactor: split admin.ts command clusters into domain modules (#76)
admin.ts (827 lines) was a flat dispatcher + all per-subcommand handlers. Pull
the two largest clusters out, behaviour-preserving:

- admin-shared.ts — request/result types, logAdmin, extractUserId/extractChannelId.
  A shared base so handler modules don't cycle back through admin.ts.
- admin-review.ts — the `review …` subcommand (provider/strategy/approval config).
- admin-audience.ts — the `audience …` subcommand (per-user/channel/default tiers
  + the v0.15 workspace examples).

admin.ts keeps handleAdminSlash + status/escalation/atoms/admins/audit/doctor and
re-exports extractUserId/extractChannelId/RuntimeHealthSnapshot for back-compat
(slash-command.ts + tests import them from here). admin.ts 827 → 445 lines.

Behaviour-preserving: handleAdminSlash is exercised by 57 test call-sites across
gateway.test / slack-adapter.test; all pass (1009 cli tests, 0 fail), typecheck
clean. No new tests — the extraction is covered by the existing dispatch tests.
